7월 8일
관리자 / 마이페이지 - 주문상세
- 주문상세 페이지 완성
- 모든 데이터가 들어가 있으며, 계산이 맞음
- 이미지는 제외시킴
- 관리자페이지는 그대로 진행, 마이페이지의 상세조회는 css 손 볼 필요가 있음

order.jsp페이지

포인트를 상품가격을 초과하여 구매를 할 수 있는 문제점

함수 return을 배열로 리턴하여, 상품가격과 배송비를 받아옴

상품가격과 배송비를 더한 값을 초과하지 않도록 inputvalue의 값을 변경

사용 포인트에 따라 포인트 적립을 변동적으로 하려했으나,
주문 상세 페이지에서 사용한 포인트, 적립한 포인트에 문제가 생겨
포인트 사용 여부에 상관없이 상품금액의 0.05%로 하기로 결정
