+++ project/src/main/java/com/myport/controller/MyPageController.java
@@ -9,6 +9,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -82,6 +83,18 @@ public class MyPageController {
 
 		return "/mypage/ordersHistoryList";
 	}
+	
+	// 주문 상세 페이지
+		@GetMapping({ "/orderInfo"})
+		public String orderInfoGET(String orderId, @ModelAttribute("cri") Criteria cri, Model model) throws Exception {
+			log.info("orderInfoGET..." + orderId);
+
+			// 선택 작가 정보
+			model.addAttribute("orderInfo", mypageService.gerOrderInfo(orderId));
+			
+			return "/mypage/orderInfo";
+
+		}
 
 	// 주문삭제
 	@PostMapping("/orderCancle")
