+++ project/src/main/java/com/myport/domain/OrderInfoDTO.java
@@ -54,5 +54,6 @@ public class OrderInfoDTO {
 	// 주문 수량
 	private int bookCount;
 	
-	
+	// 책 ID
+	private int bookId;
 }
