+++ project/src/main/java/com/myport/mapper/AdminMapper.java
@@ -50,9 +50,6 @@ public interface AdminMapper {
 	// 주문 상품 리스트
 	public List<OrderDTO> getOrderList(Criteria cri);
 	
-	// 주문 상세
-	public List<OrderInfoDTO> gerOrderInfo(String orderId);
-	
 	// 주문 총 갯수
 	public int getOrderTotal(Criteria cri);
