+++ project/src/main/java/com/myport/service/AdminService.java
@@ -41,9 +41,6 @@ public interface AdminService {
 	// 주문 상품 리스트
 	public List<OrderDTO> getOrderList(Criteria cri);
 	
-	// 주문 상세
-	public List<OrderInfoDTO> gerOrderInfo(String orderId);
-	
 	// 주문 총 갯수
 	public int getOrderTotal(Criteria cri);
