+++ project/src/main/resources/com/myport/mapper/MyPageMapper.xml
@@ -23,6 +23,17 @@
 		]]>
 	</select>
 	
+	<!-- 주문 상세 -->
+	<select id="gerOrderInfo" resultType="com.myport.domain.OrderInfoDTO">
+	<![CDATA[
+		select a.orderid, a.addressee, a.memberaddr1, a.memberaddr2, a.memberaddr3,
+		a.usepoint, b.savepoint, c.bookname, b.bookCount,b.bookPrice, b.bookDiscount ,a.deliverycost,a.orderstate, a.orderdate, c.bookId
+		from book_order a, book_orderitem b, book_product c
+		where a.orderid = b.orderid and b.bookid = c.bookid
+		and a.orderId = #{orderId}
+		]]>
+	</select>
+	
 	<!-- 주문상품 물품이름 출력 -->
 	<select id="getMyOrderProductName" resultType="com.myport.domain.BookVO">
 		select bookName from book_product a, book_order b, book_orderitem c
