+++ project/src/test/java/com/myport/mapper/MypageMapperTests.java
@@ -9,6 +9,7 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import com.myport.domain.BookVO;
+import com.myport.domain.OrderInfoDTO;
 
 import lombok.extern.log4j.Log4j;
 
@@ -28,6 +29,15 @@ public class MypageMapperTests {
 //		for(BookVO s : list) {
 //			System.out.println(s.getBookName());
 //		}
+		
+	}
+	
+	@Test
+	public void orderInfo() {
+		
+		List<OrderInfoDTO> list = mapper.gerOrderInfo("admin_2022070746");
+		
+		System.out.println(list);
 		
 	}
 }
